+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - `UIComponent.whenUndockedOnce()` and `whenDockedOnce` callbacks
 - `onTabSelected` callback in UIComponent when connected to a TabPane
 - `finally(callback)` for runAsync and other task builders
+- `Window.aboutToBeShown` property avoid false positives for invisible Workspace warning (https://github.com/edvin/tornadofx/issues/755)
 
 ## [1.7.16]